requirements.txt
flask
gunicorn
google-genai
python-dotenv
Pillow
requests
boto3
opencv-contrib-python-headless
numpy
